README
FACE MASK DETECTION

Perhaps one of the most striking lifestyle changes resulting from the COVID-19 pandemic is the mandatory use of face masks in grocery stores, restaurants and other public places. Wearing a mask, especially when in close proximity to others, is imperative to slowing the spread of COVID-19. Scientists have proven that masking lowers COVID-19 cases.
But we are finding people in public who are not wearing their mask properly. This decreases the effectiveness of masking and increases the wearer’s risk of catching and spreading the disease.

OBJECTIVE: Create a real-time face mask detector to check whether a person in wearing a mask properly or not.

IN THE PROJECT:
	Convolutional neural networks is used to train the model. For real-time detection, open-cv is used.

FUTURE SCOPE:
	This model can be implemented in cameras in public places. Already this technology has been put to use and are in the beginning stages. This can be developed in to a software which would give warning buzzers when someone is found without a mask or not following social distancing. 

STEPS TO USE : 
    Just run the code and you are good to go.
    Also the file 'mymodel.h5' which is the model trained by is attached in the repository. Just download and load the model to save computation and time! Cheers!